feat: add payment confirmation rules, email responses for declined & aproved orders and payments

+++ src/main/java/leepans/dto/cupomDesconto/CupomDescontoResponseDTO.java
@@ -1,13 +1,14 @@
 package leepans.dto.cupomDesconto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CupomDescontoResponseDTO(
         Long id,
         String codigo,
-        Double valorDesconto,
-        Double percentualDesconto,
-        Double valorMinimoCompra,
+        BigDecimal valorDesconto,
+        BigDecimal percentualDesconto,
+        BigDecimal valorMinimoCompra,
         LocalDateTime dataValidade,
         Integer quantidadeDisponivel,
         boolean ativo,
