+++ src/main/java/leepans/dto/pedido/ItemPedidoResponseDTO.java
@@ -1,11 +1,13 @@
 package leepans.dto.pedido;
 
+import java.math.BigDecimal;
+
 import leepans.dto.panela.PanelaEcommerceDTO;
 
 public record ItemPedidoResponseDTO(
         Long id,
         PanelaEcommerceDTO panela,
         Integer quantidade,
-        Double valorUnitario
+        BigDecimal valorUnitario
 ) {
 }
